requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
attrs==25.1.0
beautifulsoup4==4.13.3
bs4==0.0.2
cachetools==5.5.1
cartesia==1.3.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
daily-python==0.14.2
dataclasses-json==0.6.7
deepgram-sdk==3.8.0
Deprecated==1.2.18
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
faiss-cpu==1.10.0
fastapi==0.115.8
fastapi-cli==0.0.7
filetype==1.2.0
flatbuffers==25.1.24
flexcache==0.3
flexparser==0.4
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-texttospeech==2.24.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
humanfriendly==10.0
idna==3.10
iterators==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
llama-cloud==0.1.11
llama-index==0.12.15
llama-index-agent-openai==0.4.3
llama-index-cli==0.4.0
llama-index-core==0.12.15
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.4
llama-index-llms-openai==0.3.17
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.4
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-faiss==0.3.0
llama-parse==0.5.20
loguru==0.7.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
onnxruntime==1.20.1
openai==1.59.9
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
Pint==0.24.4
pipecat==0.3.0
pipecat-ai==0.0.54
platformdirs==4.3.6
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pyloudnorm==0.1.1
pyparsing==3.2.1
pypdf==5.2.0
python-dateutil==2.9.0.post0
python-deepcompare==2.1.0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
scipy==1.15.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
soxr==0.5.0.post1
SQLAlchemy==2.0.37
starlette==0.45.3
striprtf==0.0.26
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==13.1
wrapt==1.17.2
yarl==1.18.3
